Move ItemDistributorWorker class to separate file
